Almost there
//todo: endgame handling
